+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### V2.0.1
+
+* Add a feature to parse interpolation in `s-bind` directive.
+* Add a feature to parse interpolation in common attribute to link variables in template.
+
 ### V2.0.0
 
 * Add a feature to [parse templates](https://github.com/searchfe/san-eslint-parser/pull/5) in `ts/js` file.